DEV_20225_02_03_1633.

+++ DEVELOPMENT_Avril_NNAI/Framework_NNAI/NeuralNetworks/Data/Input/Input_Praise_0.cs
@@ -41,7 +41,7 @@ namespace Avril_NNAI
                     return 0;
             }
         }
-        public ulong Get_NumberOfInputValues()
+        public byte Get_NumberOfInputValues()
         {
             return _numberOfInputValues;
         }
