+++ DEVELOPMENT_Avril_NNAI/Framework_NNAI/NeuralNetworks/Data/Input/Input_Praise_1.cs
@@ -121,7 +121,7 @@ namespace Avril_NNAI
                     return 0;
             }
         }
-        public ulong Get_NumberOfInputValues()
+        public byte Get_NumberOfInputValues()
         {
             return _numberOfInputValues;
         }
